+++ src/Doka.EntityFrameworkCore.MySql/Internal/Scaffolding/MySqlDatabaseModelFactory.cs
@@ -131,8 +131,8 @@ internal sealed class MySqlDatabaseModelFactory : IDatabaseModelFactory
                 UniqueConstraintLoader.Load(pipelineContext);
                 IndexLoader.Load(pipelineContext);
                 ApplicationTimeTableLoader.Load(pipelineContext);
-                CheckConstraintLoader.Load(pipelineContext);
                 SpatialColumnLoader.Load(pipelineContext);
+                CheckConstraintLoader.Load(pipelineContext);
                 pipelineContexts.Add(pipelineContext);
             }